requirements.txt
asgiref==3.6.0
Django==4.2
django-import-export==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
django-cors-headers==4.0.0
